requirements.txt
cassandra-driver==3.29.2
click==8.1.8
dnspython==2.7.0
geomet==0.2.1.post1
mariadb==1.1.12
packaging==24.2
psycopg2==2.9.10
pymongo==4.11.2
six==1.17.0
